Persist determination whether build contains a submittable change
This is necessary to ensure we don't perform part of the build with the
assumption that it doesn't contain a submittable change while performing
a later part with the determination that it just became submittable.
This could cause incomplete results.

Additionally, performing this check earlier now, allows us to prevent
allocating a node for a phase/variant that would be skipped anyway.
I.e. don't schedule a node to execute what's practically a NOP.