Add post-submit version bump support
This allows bumping the version beyond that of the current build. This
is most useful to ensure a prerelease tag is always present in the
repository's HEAD. This in turn ensures that developers are less likely
to publish build artifacts with the identifier of a valid release while
it isn't a valid release.