[FIX] Always reload config after applying source changes
This ensures that we're always executing everything after that change
with the config file at that point. This is important for merges that
introduce the config file: trying to use the config before that time is
impossible.

It also enables changing the versioning policy in a pull request and
have it take effect in that exact same pull request.