feat(git): support for other branches in subdirectory worktrees
This adds the ability to check out a branch as a separate worktree in a
subdirectory of the main worktree. Build steps then get the ability to
commit to those.

The Groovy component for Jenkins takes care of synchronizing these extra
commits to other slaves.

The purpose of this functionality is to enable publication of build
products to a separate branch. The primary use case (only one at this
time) for this is to publish generated documentation by means of a
BitBucket-pages or GitHub-pages style branch that gets served directly
by a webserver.