fix: workaround Jenkins Git plugin bug causing wrong GIT_COMMIT
Apparently, under some conditions, that I don't understand, the
GIT_COMMIT variable does _not_ contain the commit hash of the commit it
just checked out. This behavior directly contradicts the documentation
of the Git Plugin so is a bug.

This change works around that bug by using git itself to determine the
currently checked out commit.