improvement: produce more stable build names and numbers on Artifactory
In particular pull-request builds are now identified as builds to their
target branch with the pull-request and build number combined in
artifactory's build number.

Specifically, given a build '2' on pull request '9' to branch 'master'
in pipeline 'some-repos-job' this would previously be shown as:

name: "some-repos-job :: PR-9",   number: "2"

With this commit that changes to:
name: "some-repos-job :: master", number: "PR-9 2"

This has the effect of grouping the builds that affect the same branch
in artifactory.