fix: don't forget to delete checkouts if we don't have change-only steps
Because we don't always have a change-only step we cannot use its
existence as a criterium for whether to delete a checkout. Instead we're
only allowed to use it as a criterium for _how_ to delete it.

Additionally group the checkout and delete-checkout code much closer
together to increase code locality. This allows us to avoid pinning
variants to nodes and subsequently unpinning it when required by not
pinning it when that would be a problem.