improvement: give --workspace a decent default
In order to make the command line interface more user friendly we're
giving the --workspace a default. We stick as close to the user's wishes
as possible with this default: if a --config file is specified we assume
its directory, otherwise the current working directory is the best we
can do.