feat: execute a command once for every source commit
In order to enable checks that need to run for every single commit in a
change-request we now enable doing just that. When specifying to execute
a command 'foreach: SOURCE_COMMIT' it will execute once for every commit
that is in the prepared change that's neither a merge nor merged in.

The latter two conditions prevents us from executing these commands on
code that's merged in from elsewhere (e.g. merges down from upstream).