fix: use version-policy specific defaults for the formatting of tags
The current default for the tag format is good only for Semantic
Versioning. As a result the generated tags for Caruso Versioning are in
essence truncated.

In order to address this problem this change makes the default tag
format specific to the used versioning policy. This enables having
different defaults for different policies. We then immediately use the
desired default format for Caruso Versioning.