feat: execute a command once for every autosquashed source commit
In order to restrict checks to single commits after autosquashing
changes to them this enables doing that. It accomplishes that by
autosquashing after a merge on a separate virtual "branch".

It works the same as 'foreach: SOURCE_COMMIT' with as its difference
that it uses 'foreach: AUTOSQUASHED_COMMIT' instead.

This is mostly useful for checks that require every single commit to be
correct: it doesn't force the user to perform an autosquash until
they're ready to merge their change request.