fix: expose --target-ref value as GIT_BRANCH directly
I.e. instead of trying to reconstruct the value from the list of
refspecs scheduled for submission.

The reason why the previous behavior was a problem is builds without a
change: they don't have a list of refspecs to submit. But they do have a
target branch to checkout just the same. So we're now exposing just
that.