fix(groovy): move regex evaluation to non-CPS context
Jenkins, apparently, isn't able to serialize java.util.regex.Matcher
objects. As a result using them in CPS context (default execution
context for Groovy scripts on Jenkins) is unsafe and can fail at random
intervals, depending on timing.

This change moves the evaluation of that regex to a @NonCPS function and
extracts all information it can from it which it then returns to do its
work on.