fix: only restore mtime for regular files and symlinks
Properly decode the object type of tree entries as documented for git's
index-format.

Then use that to explicitly only restore the mtime of regular files and
symlinks. Specifically this means that we're no longer attempting to
restore mtimes of gitlinks which are placeholders and meta data for
checking out submodules.

This addresses part of PIPE-312.