feat: make execution possible with 'hopic' as command
This adds the 'hopic' console script and makes it available to execute
from. We still retain the 'ci-driver' console script (on the release/0
branch only) and thus maintain backwards compatibility.

The intent of this is to allow us to recommend people to use the 'hopic'
command even if they're stuck on a release branch.