feat(config): add a bumping policy
Instead of only being able to specify which field to bump this change
allows to specify a policy to use for determining what to bump. To be
backwards compatible the 'constant' policy is used as default with the
string content of the previous 'bump' field as it's 'field' value.

Addresses #PIPE-348