fix: don't clear Acked-By on autosquashes
Previously, the commit hash that a reviewer had last reviewed and
approved had to match the final commit hash of the merge.
With this change, we now diff both the contents and the metadata
(commit message, author, date) of the last-reviewed commit hash of
each approval to the final merge, ignoring fixups and squash commits,
so that a change will not lose its approvers from any autosquashes
(a very common occurrence).

Fixes #PIPE-315