feat(config): search for specified Ivy manifest relative to $CFGDIR
When trying to determine the container image from an Ivy manifest file
we need to first find that file. Before this change we would search in
an unspecified location.

After this change we search for it relative to the directory holding the
Hopic CI config file first, falling back to the workspace directory.