fix: prevent pip from looking at the current repo while installing Hopic
This sometimes causes problems when using Hopic to build Hopic. Pip
doesn't just look at the virtualenv to see what packages are already
available, it also looks at the current directory. As a result, if we're
performing an incremental build with the exact same local as remote
version pip sometimes thinks Hopic is already installed. We would then
end up without Hopic in that virtualenv, breaking the build when we
attempt to use it.

This changes the working directory while pip is running to make sure
this problem doesn't occur.