improvement: switch to 'slim' image for Python instead of 'alpine'
Do this because the Alpine images use the 'musl' C library instead of
'glibc'. 'glibc' is a precondition of the virtual 'manylinux' platform.
As a result packages prebuilt for 'manylinux' cannot be used and need to
be recompiled at install time.

This takes a lot of time, for the 'regex' dependency. So we're switching
to the 'slim' image which is Debian based and contains 'glibc'.