fix: don't assume branch name is available
When building on Jenkins using a GitSCM source plugin, the
CHANGE_TARGET and BRANCH_NAME environment variables aren't available,
resulting in an obscure error.

This change adds a fallback to set Hopic's `--target-ref` to the local
source tree (which should only happen on non-pull-requests).
We use the GIT_COMMIT environment variable for that, which is guaranteed
to be set (by Hopic), instead of the BRANCH_NAME, which isn't set for
the GitSCM source plugin (for unknown reasons).