improvement: mock a username for the current uid inside docker with nss-wrapper
The openssh client mandates a username entry for the uid it's running
as. This presents a problem when attempting to checkout a git repository
over SSH within Docker with a uid that doesn't exist inside the
container.

This change uses libnss-wrapper to ensure the current uid has a username
entry associated with it.