fix: remove script approval requirement for reporting build status
To make Hopic as easy to use as possible it helps to reduce the amount
of script approvals as this is an external dependency. Accessing
properties of an Exception required a script approval and is therefore
not ideal. Instead the build status is now based upon the type of the
Exception, so far only the typical 'ABORTED' and 'FAILURE' build status are
known to raise an exception when it occurs. The
'FlowInterruptedException' is thrown when the build is aborted or when
the timeout of a build exceeded.

Implements: PIPE-969