feat: make Jenkins credential id for Bitbucket API calls configurable
Currently there is a multi stage fallback that in the end tries to use the hardcoded credentialId tt_service_account_creds
When using Jenkins job configuration 'checkout over SSH', this fallback id is used as the credentialId provided by the scm is overriden in this case.
To make hopic usable in a more generic way, this credentialId is now configurable. This will allow users using checkout over SSH, which will provide the advantage of atomic pushes