feat: add 'publish version' variable
Some package managers (e.g Maven) do not correctly implement SemVer 2.0.0. Maven will fail to perform range checks with 'build metadata' component because it treats the build metadata as an alphanumeric part of the version instead of ignoring it. The PUBLISH_VERSION should resolve this issue when publishing a candidate that is going to be promoted by extracting the meta data (using PURE_VERSION in this case).

Implements: PIPE-982