fix: handle signals while stopping Docker containers
When using CTRL-c to interrupt a `hopic build`, we automatically stop
the Docker container. When an additional CTRL-c is sent, this process
gets interrupted, resulting in a leak of the Docker container, as well
as a long and cryptic error.

With this change, when receiving a signal during the "docker stop",
we'll stop listening to further signals and trigger a "docker kill" to
force the container to exit (and be removed) before exiting.

Addresses: PIPE-984