fix: make ${HOME}/.local local to the current build
Otherwise we can get poor interactions between multiple phases of a
different build that depend on conflicting Python packages.

For example, given builds A and B that both contain phases X and Y, when
they get interleaved like this:
* A.X: pip install --user foobar>=2,<3
* B.X: pip install --user foobar>=4,<5
* B.Y: use foobar for something
* A.Y: use foobar for something else

With this ordering A.Y phase will _not_ get the required version of the
'foobar' package if it ran on the same node and had a shared
${HOME}/.local directory. That's why we're making sure this directory is
no longer shared between any builds even if they share a node.