improvement: add support for multiple executors on a single node
Currently, Hopic can only handle a single executor on a node; it
is assumed that `steps.node(NODE_NAME)` will always return the
same workspace.

This is not the case when multiple executors are configured for a
node, however.
In fact, Hopic links a workspace path to a node, so if two variants
run in parallel on the same node, the workspace administration gets
clobbered by the last executor.

This change attempts to remedy that by linking workspaces to nodes
and variants, instead of only nodes. This unfortunately has to be
managed 'manually', since Jenkins does not provide any way to request
a specific node+executor combination.

Addresses: PIPE-990