feat: report build status in same way as Bitbucket Jenkins plugin
When in the Jenkins Job configuration the option 'checkout to matching local branch' is enabled, the key that is used by the Bitbucket source plugin for reporting the build status to Bitbucket is different.
This results in having multiple build statuses in Bitbucket. Solution
for this problem is to use the same key as the Bitbucket branch plugin
uses to report the build status to Bitbucket. Depending on the branch
discovery setting, the key is determined differently.