improvement: git clean sub modules and sub repositories too
This passes an extra --force flag to 'git clean' to ensure directories
containing a .git entry are also recursed into and cleaned up.

Without this sub modules and sub repositories would be left untouched.
With this they're completely deleted. For sub modules and worktrees this
doesn't cause us to lose any checked in data as that's part of the
top level git repository's object store. Separate clones will have to be
completely refetched though.