fix(groovy): keep a reference to the usage entry we're updating
Instead of assuming that the last one is the correct one. It's not
guaranteed to be the last one when we have multiple executors on the
current node. In that case, when executing variant A and B on the same
node in parallel, if A doesn't finish before B starts this entry is no
longer the last entry in the list.