fix(groovy): ensure to build the same commit of the target branch on all nodes
It's possible for the target branch to move on _during_ a
(non-submitting) build. When that happens the previous implementation
would, for new workspaces/nodes, suddenly start building a merge into
the new version of the target branch. This results in building two
different sets of sources and thus say nothing about either.

Pinning on the commit hash that the build started with is the better
approach.