feat(checkout): support checking out a specific commit of the target branch
We now support checking out a specific target commit of the given target
branch (ref). This can be used to reproduce a merge that's already part
of the target branch. This can also be used to pin every workspace in a
single CI build to the same commit.

Addresses: PIPE-1043