feat: increase abbreviated commit hash' length in version number to 14 nibbles
More specifically we're increasing it to a minimum of 14 nibbles which
should be able to deal with at least a 100 million (100e6) unique git
objects. This assumes a uniformly distributed set of bits in the prefix
of the hash. That's likely but not guaranteed. So in individual cases,
when there's a collision on the prefix, this hash may still be extended.

Fixes: PIPE-1046