improvement(build): detect and complain about declared-but-missing artifacts
With this change artifacts declared for archival, fingerprinting or
JUnit interpretation will be checked to be present. If not at least one
of those is present Hopic will now raise an error about it.

This allows these errors to be produced locally, instead of only on
Jenkins. And thus it should allow earlier detection of these problems.