feat(merge): version bumping and error checking for PRs to hotfix branches
This alters the way that version bumping works on hotfix branches. These
branches are identified by the new `hotfix-branch` regex configuration
option.

On these branches everything but the pre-release component of version
numbers is held equal to the released version it splits off from (except
that the patch component is equal to that + 1).

In addition to that most of the preconditions, as described in the
"Hotfix Version Requirements" section of the documentation are enforced
at runtime.

Implements: PIPE-1059