test: work around importlib_metadata's returning of duplicates
With importlib_metadata 4.3.0 it doesn't fully read all meta data
anymore at filter time in an attempt to increase performance.
Unfortunately, in case of *.egg-info directories in the current working
directory, this filtering is imperfect and now allows duplicates to pass
through.

This works around it by relying on 'frozenset' to identify and eliminate
the duplicates.