fix(ciDriver): test MODALITY from env, not from params
Sometimes it is required to enable/disable a modality dynamically in a pipeline (for example, based on a trigger type).
Hopic obtains modality settings via "params.MODALITY", but "params" is an immutable map. On the other hand,
Jenkins creates a envvar for every param, and envvars can be changed during the pipeline. Switching "params.MODALITY"
to "env.MODALITY" will allow to dynamically enable/disable modalities.