feat(modality): make the modality commit's message dynamically configurable
Specifically this adds the ability to use the output of a command
(stdout) as the content of the produced commit message. This command is
ran in the same environment as the preceding commands used to perform
the modality change itself.

As such this can be used to make the produced commit message depend on
the contents of the change.

Implements: PIPE-1073