fix(groovy): ignore build status for has_submittable_change after submit started
users of the cidriver have access to the public method has_submittable_change, based on the result of this method, some logic might be executed. However, the result of has_submittable_change depends on the current build status, but if there is something that fails after the git push it is pretty confusing to the user.