feat: support CLICOLOR* color control environment variables
This follows the [CLICOLOR] spec as closely as possible, while allowing
the `--color` command line option to still overrule the environment.

In addition this transitively passes on the value of the `--color`
command line option to subprocesses via these environment variables.

This reverts commit 31d6c7197e291d5d3822e17e667b3bc3e87a27bc by
reintroducing this feature. This time followed up by another commit to
make this controllable in Jenkins.

[CLICOLOR]: https://bixense.com/clicolors/